changes in Duration and BitRate
closes #87

+++ fr.centralesupelec.edf.riseclipse.iec61850.scl/src/fr/centralesupelec/edf/riseclipse/iec61850/scl/BitRateInMbPerSec.java
@@ -20,7 +20,7 @@
 */
 package fr.centralesupelec.edf.riseclipse.iec61850.scl;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 /**
  * <!-- begin-user-doc -->
@@ -51,13 +51,13 @@ public interface BitRateInMbPerSec extends SclObject {
      * </p>
      * <!-- end-user-doc -->
      * @return the value of the '<em>Value</em>' attribute.
-     * @see #setValue(BigDecimal)
+     * @see #setValue(BigInteger)
      * @see fr.centralesupelec.edf.riseclipse.iec61850.scl.SclPackage#getBitRateInMbPerSec_Value()
-     * @model required="true"
+     * @model dataType="org.eclipse.emf.ecore.xml.type.UnsignedLong" required="true"
      *        extendedMetaData="kind='simple'"
      * @generated
      */
-    BigDecimal getValue();
+    BigInteger getValue();
 
     /**
      * Sets the value of the '{@link fr.centralesupelec.edf.riseclipse.iec61850.scl.BitRateInMbPerSec#getValue <em>Value</em>}' attribute.
@@ -67,11 +67,11 @@ public interface BitRateInMbPerSec extends SclObject {
      * @see #getValue()
      * @generated
      */
-    void setValue( BigDecimal value );
+    void setValue( BigInteger value );
 
     /**
      * Returns the value of the '<em><b>Multiplier</b></em>' attribute.
-     * The default value is <code>""</code>.
+     * The default value is <code>"M"</code>.
      * The literals are from the enumeration {@link fr.centralesupelec.edf.riseclipse.iec61850.scl.UnitMultiplierEnum}.
      * <!-- begin-user-doc -->
      * <p>
@@ -85,7 +85,7 @@ public interface BitRateInMbPerSec extends SclObject {
      * @see #unsetMultiplier()
      * @see #setMultiplier(UnitMultiplierEnum)
      * @see fr.centralesupelec.edf.riseclipse.iec61850.scl.SclPackage#getBitRateInMbPerSec_Multiplier()
-     * @model default="" unsettable="true"
+     * @model default="M" unsettable="true"
      * @generated
      */
     UnitMultiplierEnum getMultiplier();
@@ -128,7 +128,7 @@ public interface BitRateInMbPerSec extends SclObject {
 
     /**
      * Returns the value of the '<em><b>Unit</b></em>' attribute.
-     * The default value is <code>""</code>.
+     * The default value is <code>"b/s"</code>.
      * <!-- begin-user-doc -->
      * <p>
      * If the meaning of the '<em>Unit</em>' attribute isn't clear,
@@ -140,7 +140,7 @@ public interface BitRateInMbPerSec extends SclObject {
      * @see #unsetUnit()
      * @see #setUnit(String)
      * @see fr.centralesupelec.edf.riseclipse.iec61850.scl.SclPackage#getBitRateInMbPerSec_Unit()
-     * @model default="" unsettable="true"
+     * @model default="b/s" unsettable="true"
      * @generated
      */
     String getUnit();
